+++ tests/OutputFormatTest.php
@@ -98,8 +98,11 @@ EOT;
 
     /**
      * @test
+     *
+     * @deprecated since version 8.8.0; will be removed in version 9.0.
+     * Use `setSpaceAfterListArgumentSeparators()` to set different spacing per separator.
      */
-    public function spaceAfterListArgumentSeparatorComplex(): void
+    public function spaceAfterListArgumentSeparatorComplexDeprecated(): void
     {
         self::assertSame(
             '.main, .test {font: italic normal bold 16px/1.2 "Helvetica",	Verdana,	sans-serif;background: white;}'
@@ -110,6 +113,26 @@ EOT;
                 '/' => '',
                 ' ' => '',
             ]))
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function spaceAfterListArgumentSeparatorComplex(): void
+    {
+        self::assertSame(
+            '.main, .test {font: italic normal bold 16px/1.2 "Helvetica",	Verdana,	sans-serif;background: white;}'
+            . "\n@media screen {.main {background-size: 100% 100%;font-size: 1.3em;background-color: #fff;}}",
+            $this->document->render(
+                OutputFormat::create()
+                    ->setSpaceAfterListArgumentSeparator(' ')
+                    ->setSpaceAfterListArgumentSeparators([
+                        ',' => "\t",
+                        '/' => '',
+                        ' ' => '',
+                    ])
+            )
         );
     }